requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
line-bot-sdk==3.5.0
Pillow==10.1.0
waitress==2.1.2
python-dotenv==1.0.0